Provide an option to dynamically mix-in the `Entity` class (or a subclass) into a target class. Implemented this by adding `mixin` option to `entity.instrument_class` function, which stuffs the mixin class into the target class's `__bases__`.
This could be useful, for example, when using third-party ORM classes with Restler--you don't want to (or can't) change these classes, but you need a way to make them usable with Restler. The dynamic mixin is equivalent to:

class MyORMClass(SomeBase, Entity):
    pass

Note that if you define your ORM class this way up front, then the dynamic mixin won't be necessary.

Note also that if SomeBase is a SQLAlchemy `declarative_base` base class, it's not currently possible to create a subclass of MyORMClass that mixes in Entity. For example, this won't work:

class MyRESTfulORMClass(MyORMClass, Entity):
    pass

So there you have it--a dynamic RESTful mixin. Whether or not this is a good idea remains to be determined.