Made paths included in response relative to host instead of app.
After thinking a bit more about how URLs and paths included in the response will be used by applications (and after actually trying to use those paths in an app), made collection and member paths relative to host instead of application because I'm thinking that apps will generally want to use paths that are relative to the host, not the application.

 * Made collection_path and get_member_path return paths that are relative to the host URL
 * Removed app_url from response.request in favor of:
   * full_url (host URL, path, query string)
   * host_url
   * app_prefix (AKA script_name)
 * Made response.request.path point at request path instead of path_info

Now, if someone really needs the app URL, it can be composed from host_url and app_prefix. If path_info is needed, app_prefix can be stripped from the front of path or collection_path.